==> Pipelines/01_without_using_pipeline.ipynb <==
## In this notebook we will work on titanic dataset, without using pipeline 

#@ Importing the required libraries
import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split
from sklearn.impute import SimpleImputer             # for missing data
from sklearn.preprocessing import OneHotEncoder      # for categorical data
from sklearn.preprocessing import MinMaxScaler       # for scaling 
from sklearn.tree import DecisionTreeClassifier      # for prediction

#@ Loading and reading the dataset
data = pd.read_csv('titanic.csv')
data.head()

# removing some unnecessary columns
data.drop(columns = ['PassengerId', 'Name', 'Ticket', 'Cabin'], inplace = True)
data.head()

Our aim is that using all data, we have to find that the passenger is survived or not........

#@ doing train/test/split
X_train, X_test, y_train, y_test = train_test_split(data.drop(columns = ['Survived']),
                                                   data['Survived'],
                                                   test_size = 0.3,
                                                   random_state = 42)

X_train.head()

y_train.head()

# checking the missing values
data.isnull().sum()

Here, in our dataset there is some missing values so applying imputation we fill those missing values.

for age columns we fill it by finding the mean value
for embarked column we fill it by finding the most frequent value

#@ Applying Imputation

si_age = SimpleImputer()                                # creating an object for age column
si_embarked = SimpleImputer(strategy = 'most_frequent') # creating an object for an embarked column

X_train_age = si_age.fit_transform(X_train[['Age']])
X_train_embarked = si_embarked.fit_transform(X_train[['Embarked']])

X_test_age = si_age.transform(X_test[['Age']])
X_test_embarked = si_embarked.transform(X_test[['Embarked']])

# So, it fill the mean values where missing values contains

X_train_age

Here, some of the columns are categorical i.e.
    
    - Sex
    - Embarked

So, we will use one-hot encoding

# one hot encoding 

ohe_sex = OneHotEncoder(sparse = False, handle_unknown = 'ignore')
ohe_embarked = OneHotEncoder(sparse = False, handle_unknown = 'ignore')

X_train_sex = ohe_sex.fit_transform(X_train[['Sex']])
X_train_embarked = ohe_embarked.fit_transform(X_train_embarked)

X_test_sex = ohe_sex.transform(X_test[['Sex']])
X_test_embarked = ohe_embarked.transform(X_test_embarked)

X_train.head()

X_train_sex

X_train_embarked

#@ droping the sex, age, embarked and store remaining columns into the new dataframe
X_train_remain = X_train.drop(columns = ['Sex', 'Age', 'Embarked'])
X_test_remain = X_test.drop(columns = ['Sex', 'Age', 'Embarked'])

# Now, concatenate

X_train_transformed = np.concatenate((X_train_remain, X_train_age, X_train_sex, X_train_embarked), axis = 1)
X_test_transformed = np.concatenate((X_test_remain, X_test_age, X_test_sex, X_test_embarked), axis = 1)

X_train_transformed

X_train_transformed.shape

clf = DecisionTreeClassifier()            # creating an object for decision tree
clf.fit(X_train_transformed, y_train)

#@ Here, our model is trained and ready to predict 

#@ Predicting the data
y_predict = clf.predict(X_test_transformed)         # storing the value inside the 'y_pred'
y_predict

#@ calculating the accuracy
from sklearn.metrics import accuracy_score
accuracy_score(y_test, y_predict)

Now, if we want to predict the data from the given input by the users.

If we want to export the model for deploying we wil use 'pickle' library

import pickle
pickle.dump(ohe_sex, open('models/ohe_sex.pkl', 'wb'))
pickle.dump(ohe_embarked, open('models/ohe_embarked.pkl', 'wb'))
pickle.dump(clf, open('models/clf.pkl', 'wb'))
