==> Pipelines/03_with_using_pipeline.ipynb <==
Now, we are going to do similar thing using pipeline

#@ Importing the libraries
import numpy as np
import pandas as pd

from sklearn.model_selection import train_test_split
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.preprocessing import OneHotEncoder
from sklearn.preprocessing import MinMaxScaler
from sklearn.pipeline import Pipeline, make_pipeline
from sklearn.feature_selection import SelectKBest, chi2
from sklearn.tree import DecisionTreeClassifier

#@ Getting the datasets
data = pd.read_csv('titanic.csv')
data.head()

## plan for creating the pipeline

Here, we have missing values in the two columns i.e. Age & Embarked. So, first step for the pipeline is to:
- `impute` the missing values using 1 column transformer
- then we will send it to 2nd column transformer which task is to do `OneHotEncoding` for two columns i.e. `Sex` and `Embarked`
- Then, from the above output we will do `scaling`
- And, we will do `feature selection` for getting top features
- We will train our model using `DecisionTreeClassifier` which will give the output

# dropping unnecessary columns
data.drop(columns = ['PassengerId', 'Name', 'Ticket', 'Cabin'], inplace = True)

# doing train/test/split
X_train, X_test, y_train, y_test = train_test_split(data.drop(columns = ['Survived']),
                                                   data['Survived'],
                                                   test_size = 0.2,
                                                   random_state = 42)

X_train.head()

y_train.sample(5)

#@ step:1 ------> ImputationTransformer
trf1 = ColumnTransformer([
    ('impute_age', SimpleImputer(), [2]),         # creating simpleimputer in numpy array
    ('impute_embarked', SimpleImputer(strategy = 'most_frequent'), [6])      # simpleimputer for embarked
], remainder = 'passthrough')

#@ step:2 ------> OneHotEncoding
trf2 = ColumnTransformer([
    ('ohe_sex_embarked', OneHotEncoder(sparse = False, handle_unknown = 'ignore'),[1,6])
], remainder = 'passthrough')

#@ step:3 ------> Scaling
trf3 = ColumnTransformer([
    ('scale', MinMaxScaler(), slice(0,10))       # we are doing feature selection so, we will be using MinMaxScaler
])

#@ step:4 ------> FeatureSelection
trf4 = SelectKBest(score_func = chi2, k = 8)

#@ step:5 ------> Training the mode
trf5 = DecisionTreeClassifier()

## Create Pipeline
Now, we are ready for creating the pipeline 

pipe = Pipeline([
    ('trf1', trf1),
    ('trf2', trf2),
    ('trf3', trf3),
    ('trf4', trf4),
    ('trf5', trf5)
])

## Pipeline V/S make_pipeline

Pipeline requires naming of steps, make_pipeline does not

(Same applies to ColumnTransformer v/s make_column_transformer)

#@ Alternative Syntx
pipe = make_pipeline(trf1, trf2, trf3, trf4, trf5)

#@ now, training the data
pipe.fit(X_train, y_train)

## Exploring the pipeline

#@ Display the pipeline
from sklearn import set_config
set_config(display = 'diagram')

#@ getting the steps
pipe.named_steps

#@ Predicting the data
y_predict = pipe.predict(X_test)

from sklearn.metrics import accuracy_score
accuracy_score(y_test, y_predict)

## Cross Validation using Pipeline

#@ cross validation using cross_val_score
from sklearn.model_selection import cross_val_score
cross_val_score(pipe, X_train, y_train, cv = 5, scoring = 'accuracy').mean()

## GridSearch using Pipeline

# gridsearchcv
params = {
    'trf5__max_depth':[1,2,3,4,5,None]
}

from sklearn.model_selection import GridSearchCV
grid = GridSearchCV(pipe, params, cv=5, scoring='accuracy')
grid.fit(X_train, y_train)

grid.best_score_

grid.best_params_

## Exporting the Pipeline

#@ Export 
import pickle 
pickle.dump(pipe, open('pipe.pkl', 'wb'))
